requirements.txt
Flask==0.12.1
Flask-SQLAlchemy==2.2
Flask-Migrate==1.3.0
Flask-Script==2.0.5
ujson==1.35
python-telegram-bot==5.3.0
